Chore : move post image url API
